int -> float

+++ orb_models/forcefield/nn_util.py
@@ -197,7 +197,7 @@ class ScalarNormalizer(torch.nn.Module):
         return x * torch.sqrt(self.bn.running_var) + self.bn.running_mean  # type: ignore
 
 
-def get_cutoff(r: torch.Tensor, r_max: int = 6) -> torch.Tensor:
+def get_cutoff(r: torch.Tensor, r_max: float = 6) -> torch.Tensor:
     """Get a hardcoded cutoff function for attention. Default cutoff is 6 angstrom."""
     p = 4  # polynomial order
     envelope = (
